refactor: create preprocessor utils
