+++ now/constants.py
@@ -5,7 +5,7 @@ from docarray.typing import Image, Text, Video
 from now.utils import BetterEnum
 
 DOCKER_BFF_PLAYGROUND_TAG = '0.0.151-demo-example-10'
-NOW_PREPROCESSOR_VERSION = '0.0.121-demo-example-10'
+NOW_PREPROCESSOR_VERSION = '0.0.121-update-preprocessor-1'
 NOW_ELASTIC_INDEXER_VERSION = '0.0.143-refactor-none-mmdoc-elastic-2'
 NOW_AUTOCOMPLETE_VERSION = '0.0.9-fix-filter-index-fields-20'
