+++ now/executor/preprocessor/executor.py
@@ -15,7 +15,7 @@ from now.executor.abstract.auth import (
     get_auth_executor_class,
     secure_request,
 )
-from now.utils import maybe_download_from_s3
+from now.executor.preprocessor.utils import maybe_download_from_s3
 
 Executor = get_auth_executor_class()
